ERB -> HAML
I'm still on the fence on this one. On the one hand, it does reveal
the structure of the document better, and I do love that the output
is so clean (compared to the messier output of ERB). However, I'm
not sure if the price is too high, or just right; it definitely
doesn't come cheap. There are some things that feel very awkward in
haml, but maybe it's just a matter of getting used to it.