"recall" prior events for easy entry of similar events
This is the "memorized" transactions feature. You enter an actor
name, click the "recall" link, and Bucketwise pulls up the last
event you entered that used that actor. Click "recall" again and
it gets the one before that, etc.

Note that, for usability purposes, this swaps the position of
the actor and amount fields. Before, you'd enter the amount and
then the actor. Now, you enter the actor and then the amount
(because you may want to recall the amount based on the actor).